Separated RequirementsFragment into it's own class
